backend completed

+++ backend/controllers/goalControllers.js
@@ -1,66 +1,88 @@
 const asyncHandler = require('express-async-handler');
 const Goal = require('../models/goalModel');
-const { text } = require('body-parser');
+const User = require('../models/userModel'); // Make sure to import the User model
+
 // @desc  GET goals
 // @route GET/api/goals
-// @acess Private
+// @access Private
 const getGoals = asyncHandler(async (req, res) => {
-  const goals = await Goal.find({});
+  const goals = await Goal.find({ user: req.user.id });
   res.status(200).json(goals);
 });
 
 // @desc  Set goals
 // @route POST/api/goals
-// @acess Private
+// @access Private
 const setGoals = asyncHandler(async (req, res) => {
   if (!req.body.text) {
-    // text is empty
-    // res.status(400).json({ mesage: 'Please add a text field' });
     res.status(400);
-    throw new Error('please add a text field');
+    throw new Error('Please add a text field');
   }
   const goal = await Goal.create({
     text: req.body.text,
+    user: req.user.id,
   });
-  console.log(req.body);
-  // res.status(200).json({ mesage: 'set goals' });
   res.status(200).json(goal);
 });
 
 // @desc  Update goals
 // @route PUT/api/goals/:id
-// @acess Private
+// @access Private
 const updateGoal = asyncHandler(async (req, res) => {
   const goal = await Goal.findById(req.params.id);
   if (!goal) {
-    res.status(400);
+    res.status(404);
     throw new Error('Goal not found');
   }
-  const updateGoals = await Goal.findByIdAndUpdate(req.params.id, req.body, {
+
+  const user = await User.findById(req.user.id); // Corrected
+
+  if (!user) {
+    res.status(401);
+    throw new Error('User not found');
+  }
+
+  if (goal.user.toString() !== user.id) {
+    res.status(401);
+    throw new Error('User not Authorized');
+  }
+
+  const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {
     new: true,
   });
-  res.status(200).json(updateGoals);
+
+  res.status(200).json(updatedGoal);
 });
 
 // @desc  Delete goals
 // @route DELETE/api/goals/:id
-// @acess Private
+// @access Private
 const deleteGoal = asyncHandler(async (req, res) => {
   const goal = await Goal.findById(req.params.id);
 
   if (!goal) {
-    res.status(404); // Not Found
+    res.status(404);
     throw new Error('Goal not found');
   }
 
-  await goal.deleteOne(); // or goal.deleteMany() if applicable
+  const user = await User.findById(req.user.id); // Corrected
+
+  if (!user) {
+    res.status(401);
+    throw new Error('User not found');
+  }
+
+  if (goal.user.toString() !== user.id) {
+    res.status(401);
+    throw new Error('User not Authorized');
+  }
+
+  await goal.deleteOne();
 
   res
     .status(200)
     .json({ id: req.params.id, message: 'Goal deleted successfully' });
 });
-
-module.exports = deleteGoal;
 
 module.exports = {
   getGoals,
